backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
groq==0.5.0
python-dotenv==1.0.1
pydantic==2.6.3
httpx==0.27.0
